Property details show on marker click finished.
