+++ src/components/Map/Marker.tsx
@@ -1,5 +1,77 @@
-import React from 'react';
+import React, {useCallback, useState} from 'react';
 import './Marker.css';
+import qs from "qs";
+import
+  axios, {
+  AxiosError,
+  AxiosResponse
+} from "axios";
+import { IPropertyData } from "../Sidebar/Sidebar";
+import PropertyDetailView from "../PropertyDetailView/PropertyDetailView";
+
+export interface PropertyQuery {
+  addressNumber: string;
+  address: string;
+  city: string;
+  state: string;
+  zipcode: string;
+}
+
+/**
+ * Configure property query for use
+ */
+const getPropertyData = (queryAddress: PropertyQuery) => {
+  const {
+    addressNumber,
+    address,
+    city,
+    state,
+    zipcode
+  } = queryAddress;
+
+  /**
+   * Authorization header for the Particle Space API
+   */
+  const data = qs.stringify({
+    'secret_key': process.env.REACT_APP_PARTICLE_SPACE_SECRET_KEY,
+    'publish_key': process.env.REACT_APP_PARTICLE_SPACE_PUBLISH_KEY,
+  });
+  const authorizationConfig = {
+    method: 'post',
+    url: 'https://api.particlespace.com/api/v1/authenticate',
+    headers: {
+      'Content-Type': 'application/x-www-form-urlencoded'
+    },
+    data : data
+  };
+
+  axios(authorizationConfig)
+    .then(function (response: AxiosResponse) {
+      console.log(JSON.stringify(response.data));
+    })
+    .catch(function (error: AxiosError) {
+      console.log(error);
+    });
+
+  /**
+   * Particle Space search API
+   */
+  const searchConfig = {
+    method: 'get',
+    url: `https://api.particlespace.com/api/v1/property/search?address=${addressNumber} ${address}&city=${city}&state=${state}&zipcode=${zipcode}`,
+    headers: {
+      'Authorization': 'Bearer ' + process.env.BEARER_TOKEN
+    }
+  };
+
+  axios(searchConfig)
+    .then(function (response: AxiosResponse) {
+      console.log(JSON.stringify(response.data));
+    })
+    .catch(function (error: AxiosError) {
+      console.log(error);
+    });
+}
 
 export interface IMarker {
   color: string;
@@ -8,7 +80,7 @@ export interface IMarker {
   id: string | number;
   lat: number;
   lng: number;
-  onClick?: () => void;
+  propertyData: IPropertyData;
 }
 
 export const Marker = (props: IMarker) => {
@@ -16,7 +88,11 @@ export const Marker = (props: IMarker) => {
     color,
     name,
     text,
+    propertyData
   } = props;
+
+  const [isOpen, setOpen] = useState(false);
+
   const handleMouseOver = (event: React.MouseEvent<HTMLDivElement>) => {
     const box: HTMLDivElement = event.currentTarget;
       box.style.backgroundColor = "#000a3d";
@@ -25,10 +101,19 @@ export const Marker = (props: IMarker) => {
     const box: HTMLDivElement = event.currentTarget;
       box.style.backgroundColor = "#0078ff";
   };
+  const handleOnClick = useCallback(() => {
+    setOpen(true);
+    console.log(isOpen)
+  }, [setOpen]);
 
   return (
     <div>
-      <div className="marker bounce"
+      <PropertyDetailView
+        isOpen={isOpen}
+        setOpen={setOpen}
+        propertyData={propertyData}
+      />
+        <div className="marker bounce"
         style={{
           backgroundColor: color,
           cursor: 'pointer'
@@ -36,7 +121,7 @@ export const Marker = (props: IMarker) => {
         title={name}
         onMouseEnter={handleMouseOver}
         onMouseLeave={handleMouseLeave}
-        onClick={props.onClick}
+        onClick={handleOnClick}
       />
       <div className='textbox'>
         {text}
